requirement.txt
opencv-python
picamera2
adafruit-circuitpython-servokit
pyttsx3
face_recognition
RPi.GPIO
